requirements.txt
alembic==1.0.11
amqp==2.5.0
aniso8601==7.0.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
Click==7.0
coverage==4.5.4
filelock==3.0.12
Flask==1.1.1
Flask-JWT-Extended==3.21.0
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
importlib-metadata==0.19
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
lazy-object-proxy==1.4.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.20.1
marshmallow-sqlalchemy==0.17.0
mccabe==0.6.1
meld3==1.0.2
more-itertools==7.2.0
packaging==19.1
passlib==1.7.1
pep8==1.7.1
pluggy==0.12.0
psycopg2-binary==2.8.3
py==1.8.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
pytest==5.1.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.6
SQLAlchemy-Utils==0.34.1
supervisor==4.0.4
toml==0.10.0
tox==3.13.2
typed-ast==1.4.0
vine==1.3.0
virtualenv==16.7.2
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
yapf==0.28.0
zipp==0.5.2
